Move event handling to it's own class
