+++ Browser/browser.py
@@ -11,6 +11,7 @@ from .keywords import (
     Cookie,
     Devices,
     Evaluation,
+    EventHandling,
     Getters,
     Interaction,
     Network,
@@ -555,6 +556,7 @@ class Browser(DynamicCore):
             Control(self),
             Cookie(self),
             Devices(self),
+            EventHandling(self),
             Evaluation(self),
             Interaction(self),
             Getters(self),
