+++ Browser/keywords/interaction.py
@@ -1,5 +1,4 @@
 import json
-from pathlib import Path
 from typing import Dict, Optional
 
 from robotlibcore import keyword  # type: ignore
@@ -8,7 +7,6 @@ from ..base import LibraryComponent
 from ..generated.playwright_pb2 import Request
 from ..utils import logger
 from ..utils.data_types import (
-    AlertAction,
     BoundingBox,
     Coordinates,
     KeyAction,
@@ -326,43 +324,6 @@ class Interaction(LibraryComponent):
             )
             if log_response:
                 logger.debug(response.log)
-
-    @keyword(tags=["Setter", "PageContent", "EventHandler"])
-    def upload_file(self, path: str):
-        """ Upload file from ``path`` into next file chooser dialog on page.
-
-        ``path`` <str> Path to file to be uploaded.
-
-        Example use:
-
-        | Upload File    ${CURDIR}/test_upload_file
-        | Click          \\#file_chooser
-
-        """
-        p = Path(path)
-        p.resolve(strict=True)
-        with self.playwright.grpc_channel() as stub:
-            response = stub.UploadFile(Request().FilePath(path=str(p)))
-            logger.debug(response.log)
-
-    @keyword(tags=["PageContent", "EventHandler"])
-    def handle_alert(self, action: AlertAction, prompt_input: str = ""):
-        """ Handle next dialog on page with ``action``. Dialog can be any of alert,
-        beforeunload, confirm or prompt.
-
-            ``action`` < ``acceppt`` | ``dismiss`` > How to handle the alert. **Required**
-
-            ``prompt_input`` <str> The value to enter into prompt. Only valid if
-            ``action`` equals accept. Defaults to empty string.
-        """
-
-        with self.playwright.grpc_channel() as stub:
-            if prompt_input and action is not AlertAction.accept:
-                raise ValueError("prompt_input is only valid if action is 'accept'")
-            response = stub.HandleAlert(
-                Request().AlertAction(alertAction=action.name, promptInput=prompt_input)
-            )
-            logger.debug(response.log)
 
     @keyword(tags=["VirtualMouse", "PageContent"])
     def mouse_button(
